fixed security issues with the delete handler

+++ engsci_conga/File_handling/views.py
@@ -30,7 +30,7 @@ def delete_handler(request, user, pk):
     if f.owner != request.user:
         error = u"This is not your file! You are now on Santa's naughty list!!"
 
-    if request.method == 'POST':
+    if request.method == 'POST' and error is not None:
         f.delete()
         return HttpResponseRedirect(reverse(
             'engsci_conga.views.courses_view',
